requirements.txt
alembic==1.8.1
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
azure-ai-formrecognizer==3.2.0
azure-common==1.1.28
azure-core==1.26.0
backoff==1.11.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudpickle==2.2.0
databricks-cli==0.17.3
Deprecated==1.2.13
dill==0.3.6
Django==4.1.3
djangorestframework==3.14.0
docker==6.0.1
docopt==0.6.2
elasticsearch==7.17.7
entrypoints==0.4
farm-haystack==1.10.0
filelock==3.8.0
Flask==2.2.2
gitdb==4.0.9
GitPython==3.1.29
greenlet==2.0.0
gunicorn==20.1.0
huggingface-hub==0.10.1
idna==3.4
importlib-metadata==5.0.0
inflect==6.0.2
isodate==0.6.1
itsdangerous==2.1.2
jarowinkler==1.2.3
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.0
kombu==5.2.4
langdetect==1.0.9
lxml==4.9.1
Mako==1.2.3
MarkupSafe==2.1.1
mlflow==1.30.0
mmh3==3.0.0
monotonic==1.6
more-itertools==9.0.0
msrest==0.7.1
networkx==2.8.8
nltk==3.7
num2words==0.5.12
numpy==1.23.4
oauthlib==3.2.2
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
posthog==2.1.2
prometheus-client==0.15.0
prometheus-flask-exporter==0.20.3
prompt-toolkit==3.0.31
protobuf==4.21.9
pydantic==1.10.2
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.6
PyYAML==6.0
quantulum3==0.7.11
querystring-parser==1.2.4
rapidfuzz==2.7.0
redis==4.3.4
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
scikit-learn==1.1.3
scipy==1.9.3
sentence-transformers==2.2.2
sentencepiece==0.1.97
seqeval==1.2.2
six==1.16.0
smmap==5.0.0
SQLAlchemy==1.4.42
sqlparse==0.4.3
tabulate==0.9.0
threadpoolctl==3.1.0
tika==1.24
tokenizers==0.12.1
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.1
transformers==4.21.2
typing_extensions==4.4.0
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
websocket-client==1.4.1
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.10.0
